+++ Changelog.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `VK_KHR_sampler_ycbcr_conversion` device extension (#785)
 - Added `VK_EXT_swapchain_maintenance1` device extension (#786)
 - Added `VK_NV_low_latency2` device extension (#802)
+- Added `VK_EXT_hdr_metadata` device extension (#804)
 
 ### Changed